feat: Update message system / Fix defunct problem

+++ np_multi_proc_dir/src/np_multi_proc.c
@@ -2,6 +2,7 @@
 #include <stdlib.h>
 #include <unistd.h>
 #include <signal.h>
+#include <errno.h>
 #include <sys/socket.h>
 #include <sys/wait.h>
 #include <netinet/in.h>
@@ -9,10 +10,15 @@
 #include "npshell.h"
 #include "user.h"
 #include "prompt.h"
+#include "pidlist.h"
+
+static pid_list *child_plist;
 
 static void np_multi_proc_init(void);
 
 static void SIGINT_handler(int signum);
+static void SIGINT_child_handler(int signum);
+static void SIGCHLD_handler(int signum);
 
 // ./np_multi_proc 12345 # Listen on 0.0.0.0:12345
 int main(int argc, char **argv)
@@ -73,6 +79,7 @@ int main(int argc, char **argv)
 
         if ((pid = fork()) > 0) {
             // Parent process
+            plist_insert_block(child_plist, pid);
 
             // Close unused fd
             close(cs);
@@ -81,11 +88,25 @@ int main(int argc, char **argv)
             char msgbuf[0x60] = { 0 };
             uint32_t uid;
 
+            // New session
+            if (setsid() < 0) {
+                perror("setsid()");
+                exit(errno);
+            }
+
+            signal(SIGINT, SIGINT_child_handler);
+
+            // printf("[>] -- login\n");
+
             // New user
             uid = user_new(caddr);
+            
+            // printf("[>] %d login\n", uid);
 
             // Welcome message
             msg_motd(cs);
+
+            // printf("[>] %d login motd\n", uid);
 
             // Broadcast login message
             sprintf(msgbuf, "*** User '%s' entered from %s:%d. ***\n", 
@@ -95,6 +116,8 @@ int main(int argc, char **argv)
 
             user_broadcast(msgbuf, MSG_NONE);
             msg_tell(cs, msgbuf);
+
+            // printf("[>] %d login user_broadcast & tell\n", uid);
 
             msg_reset_read_offset();
 
@@ -108,13 +131,23 @@ int main(int argc, char **argv)
             sprintf(msgbuf, "*** User '%s' left. ***\n", \
                 user_manager.all_users[uid].name);
 
+            // printf("[>] %d logout\n", uid);
+
             user_broadcast(msgbuf, MSG_LOGOUT);
 
+            // printf("[>] %d logout broadcast\n", uid);
+
             user_release(uid);
+
+            // printf("[>] %d logout release\n", uid);
+
+            // Kill all child command
+            kill(0, SIGKILL);
 
             exit(0);
         } else {
             // TODO: Handle fork error
+            perror("[x] fork error\n");
         }
     }
 }
@@ -122,10 +155,13 @@ int main(int argc, char **argv)
 static void np_multi_proc_init(void)
 {
     signal(SIGINT, SIGINT_handler);
+    signal(SIGCHLD, SIGCHLD_handler);
     signal(SIGUSR1, SIG_IGN);
 
     user_manager_init();
     global_msg_init();
+    
+    child_plist = plist_init();
 }
 
 static void SIGINT_handler(int signum)
@@ -135,14 +171,42 @@ static void SIGINT_handler(int signum)
 
     // Exit
 
+    // Close all shell
+    for (pid_node *pn = child_plist->head; pn; pn = pn->next) {
+        kill(pn->pid, SIGINT);
+    }
+
     user_manager_release();
     global_msg_release();    
-
-    // Close all shell
-    kill(0, SIGKILL);
 
     // TODO: Close all FIFO
 
     // Bye
     exit(0);
+}
+
+static void SIGINT_child_handler(int signum)
+{
+    if (signum != SIGINT)
+        return;
+
+    user_release(global_uid);
+    
+    // Kill all child command
+    kill(0, SIGKILL);
+
+    // Bye
+    // exit(0);
+}
+
+static void SIGCHLD_handler(int signum)
+{
+    int cpid;
+
+    if (signum != SIGCHLD)
+        return;
+
+    cpid = wait(NULL);
+
+    plist_delete_by_pid(child_plist, cpid);
 }
